requirements.txt
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
dataclasses==0.6
future==0.18.2
idna==2.10
kiwisolver==1.3.1
matplotlib==3.2.0
numpy==1.19.3
pandas==1.1.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-nlp==0.5.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
sentencepiece==0.1.94
six==1.15.0
torch==1.7.0+cu110
torchaudio==0.7.0
torchtext==0.6.0
torchvision==0.8.1+cu110
tqdm==4.51.0
typing-extensions==3.7.4.3
urllib3==1.25.11
